+++ src/Products/CMFPlomino/contents/form.py
@@ -8,6 +8,7 @@ from plone import api
 from plone.app.z3cform.wysiwyg import WysiwygFieldWidget
 from plone.autoform import directives as form
 from plone.dexterity.content import Container
+from plone.memoize import instance
 from plone.supermodel import directives, model
 from pyquery import PyQuery as pq
 from z3c.form.datamanager import AttributeField, zope
@@ -314,18 +315,27 @@ class PlominoForm(Container):
         return value
 
     def getIsMulti(self):
-        # TODO: Cache this?
+        db = self.getParentDatabase()
+        cache_key = 'getIsMulti_%s'%repr(self)
+
+        cached = db.getRequestCache(cache_key)
+        if cached is not None:
+            return cached
+
         html_content = self.form_layout or ''
         if not html_content:
             return False
 
         # The form is multipage if there is a pagebreak HR
         html = pq(html_content)
+        result = False
         for elem in html.children():
             if elem.tag == 'hr' and \
                             elem.attrib.get('class') == 'plominoPagebreakClass':
-                return True
-        return False
+                result = True
+                break
+        db.setRequestCache(cache_key, result)
+        return result
 
     def getFormAction(self, request, bare=False):
         """
@@ -518,17 +528,18 @@ class PlominoForm(Container):
         """
         db = self.getParentDatabase()
         # XXX: Can this be re-enabled?
-        # cache_key = "getFormFields_%d_%d_%d_%d_%d_%d" % (
-        #     hash(self),
-        #     hash(doc),
-        #     includesubforms,
-        #     applyhidewhen,
-        #     validation_mode,
-        #     deduplicate
-        # )
-        # cache = db.getRequestCache(cache_key)
-        # if cache:
-        #     return cache
+        cache_key = "getFormFields_%s_%d_%d_%d_%d_%d" % (
+            repr(self),
+            hash(doc),
+            #hash(frozenset(request.keys())),
+            includesubforms,
+            applyhidewhen,
+            validation_mode,
+            deduplicate
+        )
+        cache = db.getRequestCache(cache_key)
+        if cache:
+            return cache
         # if not request and hasattr(self, 'REQUEST'):
         #     request = self.REQUEST
         form = self.getForm()
@@ -582,7 +593,7 @@ class PlominoForm(Container):
                      for f, c in seen.items() if c > 1])
                 logger.debug('Overridden fields: %s' % report)
 
-        # db.setRequestCache(cache_key, result)
+        db.setRequestCache(cache_key, result)
         return result
 
     security.declarePublic('getHidewhenFormulas')
@@ -601,7 +612,11 @@ class PlominoForm(Container):
     def getHidewhen(self, id):
         """ Get a hide-when formula
         """
-        return getattr(self, id)
+        hidewhen = getattr(self, id)
+        if getattr(hidewhen, 'portal_type', '') != 'PlominoHidewhen':
+            raise AttributeError()
+        return hidewhen
+
 
     def getFormActions(self):
         """ Get all actions
@@ -822,7 +837,7 @@ class PlominoForm(Container):
 
     @plomino_profiler('form')
     def displayDocument(self, doc, editmode=False, creation=False,
-                        parent_form_id=False, request=None):
+            parent_form_id=False, request=None, temp_doc=None):
         """ Display the document using the form's layout
         """
         db = self.getParentDatabase()
@@ -836,22 +851,35 @@ class PlominoForm(Container):
             if form_mode and form_mode == 'WRITE':
                 editmode = True
 
-        if parent_form_id:
-            parent_form = db.getForm(parent_form_id)
+        if request:
+            parent_form_ids = db.getRequestCache('parent_form_ids')
+            if parent_form_ids is None:
+                parent_form_ids = []
+            if parent_form_id:
+                db.setRequestCache('parent_form_ids', parent_form_ids + [parent_form_id])
+        else:
+            parent_form_ids = []
+
+        if parent_form_ids:
+            parent_form = db.getForm(parent_form_ids[0])
             is_multi = parent_form.getIsMulti()
         else:
             is_multi = self.getIsMulti()
 
         # Use the request for the temp doc in editmode with a multipage form
         if request and editmode and request.get('form') and is_multi:
+            #TODO: why? need to work out test for this
             use_request = True
             # Use the parent form for the temp doc
             if parent_form_id:
                 tempdoc_form = parent_form
         else:
             use_request = False
+
+        if temp_doc is not None:
+            hidewhen_target = temp_doc
         # remove the hidden content
-        if doc and use_request:
+        elif doc and use_request:
             hidewhen_target = getTemporaryDocument(
                 db,
                 tempdoc_form,
@@ -869,11 +897,13 @@ class PlominoForm(Container):
             hidewhen_target = doc
         html_content = self.applyHideWhen(hidewhen_target, silent_error=False, cache_key='displayDocument')
         hidden_fields, reset_fields = self._get_hidden_fields(request, hidewhen_target, validation_mode=False)
-        if request:
-            parent_form_ids = request.get('parent_form_ids', [])
-            if parent_form_id:
-                parent_form_ids.append(parent_form_id)
-                request.set('parent_form_ids', parent_form_ids)
+        # TODO: work out why this is here. We shouldn't be changing the request as it
+        # screws up request caching
+        # if request:
+        #     parent_form_ids = request.get('parent_form_ids', [])
+        #     if parent_form_id:
+        #         parent_form_ids.append(parent_form_id)
+        #         request.set('parent_form_ids', parent_form_ids)
 
         # get the field lists
         fields = self.getFormFields(doc=doc, request=request)
@@ -901,7 +931,7 @@ class PlominoForm(Container):
                             asUnicode(request.get(field_id, '')),
                             html_content)
                     )
-            for f in self.getFormFields():
+            for f in fields:
                 if f.field_type == 'ATTACHMENT' and db.getRequestCache(f.id+"@@ATTACHMENT"):
                     temp_files = db.getRequestCache(f.id+"@@ATTACHMENT")
                     html_content = (
@@ -942,12 +972,12 @@ class PlominoForm(Container):
                 )
 
         # insert subforms
-        for subformname in self.getSubforms(doc):
+        for subformname in self.getSubforms(hidewhen_target):
             subform = self.getParentDatabase().getForm(subformname)
             if subform:
                 subformrendering = subform.displayDocument(
                     doc, editmode, creation, parent_form_id=self.id,
-                    request=request)
+                    request=request, temp_doc=hidewhen_target)
                 html_content = html_content.replace(
                     '<span class="plominoSubformClass">%s</span>' %
                     subformname,
@@ -1050,16 +1080,16 @@ class PlominoForm(Container):
         # Cache the number of pages so we don't have to parse the HTML
         # more than once per request
         db = self.getParentDatabase()
-        cache_key = 'cached-num-pages'
+        cache_key = 'cached-num-pages_%s'%repr(self)
 
         cached = db.getRequestCache(cache_key)
         if cached is not None:
             return cached
 
         pages = 0
-        html_content = self._get_html_content()
-        if html_content:
-            html = pq(html_content)
+        html = self._get_html_content(True)
+        if html:
+            #html = pq(html_content)
             pages = html('.multipage').size()
 
         # Cache the result
@@ -1215,9 +1245,19 @@ class PlominoForm(Container):
             pq(e).replace_with(span)
         self.form_layout = tostring_innerhtml(root)
 
+        #to ensure we don't get write on read with getIsMulti
+        self.getIsMulti()
+
     security.declarePrivate('_get_html_content')
 
-    def _get_html_content(self):
+    def _get_html_content(self, as_pyquery=False):
+        db = self.getParentDatabase()
+        cache_key = "_get_html_content_%s"%repr(self)
+        html = db.getRequestCache(cache_key)
+        if html is not None:
+
+            return html if as_pyquery else html.outerHtml()
+
         # get the raw value for rendering the form
         html_content = self.form_layout or ''
         html_content = html_content.replace('\r\n', '')
@@ -1275,11 +1315,11 @@ class PlominoForm(Container):
             # Remove the old node
             pq(linkto_node).remove()
 
+        db.setRequestCache(cache_key, html)
+
         # If this is html() and there is only one element, it will only return
         # what's inside it.
-        html_content = html.outerHtml()
-
-        return html_content
+        return html if as_pyquery else html.outerHtml()
 
     def example_widget(self, widget_type=None, id=None):
         """ id=None means we are dragging so just need a basic example
@@ -1389,8 +1429,8 @@ class PlominoForm(Container):
         # Some methods might pass in their own cache_key. Add the rest
         # of the key values to the key.
         db = self.getParentDatabase()
-        cache_key = "applyHideWhen_%d_%d_%d_%s" % (
-            hash(self),
+        cache_key = "applyHideWhen_%s_%d_%d_%s" % (
+            repr(self),
             hash(doc),
             hash(split_multipage),
             cache_key
@@ -1401,32 +1441,41 @@ class PlominoForm(Container):
 
         html_content = self._get_html_content()
 
-        # remove the hidden content
-        for hidewhen in self.getHidewhenFormulas():
-            hidewhenName = hidewhen.id
+        if doc is None:
+            target = self
+        else:
+            target = doc
+        for hidewhenName, result, isReset in self._get_hidewhens(target):
             try:
-                if doc is None:
-                    target = self
-                else:
-                    target = doc
-
-                result = self.runFormulaScript(
-                    SCRIPT_ID_DELIMITER.join(
-                        ['hidewhen', self.id, hidewhen.id, 'formula']),
-                    target,
-                    hidewhen.formula)
-            except PlominoScriptException, e:
-                if not silent_error:
-                    # applyHideWhen is called by getFormFields and
-                    # getSubForms; in those cases, error reporting
-                    # is not accurate,
-                    # we only need error reporting when actually rendering a
-                    # page
-                    e.reportError(
-                        '%s hide-when formula failed' % hidewhen.id,
-                        request=getattr(self, 'REQUEST', None))
-                # if error, we hide anyway
-                result = True
+                hidewhen = self.getHidewhen(hidewhenName)
+            except AttributeError:
+                # Will be multipage one
+                continue
+            #TODO: take error error reporting out of _get_hidewhens
+        #
+        #
+        # # remove the hidden content
+        # for hidewhen in self.getHidewhenFormulas():
+        #     hidewhenName = hidewhen.id
+        #     try:
+        #
+        #         result = self.runFormulaScript(
+        #             SCRIPT_ID_DELIMITER.join(
+        #                 ['hidewhen', self.id, hidewhen.id, 'formula']),
+        #             target,
+        #             hidewhen.formula)
+        #     except PlominoScriptException, e:
+        #         if not silent_error:
+        #             # applyHideWhen is called by getFormFields and
+        #             # getSubForms; in those cases, error reporting
+        #             # is not accurate,
+        #             # we only need error reporting when actually rendering a
+        #             # page
+        #             e.reportError(
+        #                 '%s hide-when formula failed' % hidewhen.id,
+        #                 request=getattr(self, 'REQUEST', None))
+        #         # if error, we hide anyway
+        #         result = True
 
             start = ('<span class="plominoHidewhenClass">start:%s</span>' %
                      hidewhenName)
@@ -1550,10 +1599,28 @@ class PlominoForm(Container):
 
         # Only evaluate hidewhens if there is something to evaluate
         hidewhens = asList(REQUEST.get('_hidewhens', []))
-        if hidewhens:
-            results['hidewhens'] = self._get_hidewhens(REQUEST, doc, temp=temp, dynamic=True)
-        else:
-            results['hidewhens'] = []
+        results['hidewhens'] = []
+        for token in hidewhens:
+            (formid, hwid) = token.split('/')
+            if formid == self.id:
+                form = self
+            else:
+                form = db.getForm(formid)
+                if not form:
+                    db.writeMessageOnPage(
+                        "Form %s id missing" % formid, REQUEST, False)
+                    results['hidewhens'].append(["%s/%s" % (formid, hwid), True])
+                    continue
+            if formid not in temp:
+                temp[formid] = getTemporaryDocument(
+                    db,
+                    db.getForm(formid),
+                    REQUEST,
+                    doc,
+                    validation_mode=False)
+            hidewhen = form.getHidewhen(hwid)
+            _, isHidden, isReset = form._get_hidewhen(temp[formid], hidewhen)
+            results['hidewhens'].append(["%s/%s" % (formid, hwid), isHidden, isReset])
 
         fields = asList(REQUEST.get('_fields', []))
         fields_results = []
@@ -1593,80 +1660,29 @@ class PlominoForm(Container):
 
     security.declarePrivate('_get_hidewhens')
 
-    def _get_hidewhens(self, REQUEST, doc, temp=None, include_multipage=False, dynamic=False):
+    def _get_hidewhens(self, doc):
         db = self.getParentDatabase()
 
         # assume the request doesn't change and temp is based on request
-        cache_key = "_get_hidewhens_%d_%d_%d_%d" % (
-            hash(self),
+        cache_key = "_get_hidewhens_%s_%d" % (
+            repr(self),
             hash(doc),
-            hash(include_multipage),
-            hash(dynamic)
         )
         cache = db.getRequestCache(cache_key)
         if cache is not None:
             return cache
 
-        if temp is None:
-            temp = {}
         hidewhens_results = []
 
-        if dynamic:
-            hidewhens = asList(REQUEST.get('_hidewhens', []))
-            for token in hidewhens:
-                (formid, hwid) = token.split('/')
-                if formid == self.id:
-                    form = self
-                else:
-                    form = db.getForm(formid)
-                    if not form:
-                        db.writeMessageOnPage(
-                            "Form %s id missing" % formid, REQUEST, False)
-                        hidewhens_results.append(["%s/%s" % (formid, hwid), True])
-                        continue
-                if formid not in temp:
-                    temp[formid] = getTemporaryDocument(
-                        db,
-                        db.getForm(formid),
-                        REQUEST,
-                        doc,
-                        validation_mode=False)
-                try:
-                    hidewhen = form.getHidewhen(hwid)
-                    isHidden = form.runFormulaScript(
-                        SCRIPT_ID_DELIMITER.join([
-                            'hidewhen', formid, hwid, 'formula'
-                        ]),
-                        temp[formid],
-                        hidewhen.formula)
-                except PlominoScriptException, e:
-                    e.reportError(
-                        '%s hide-when formula failed' % hwid)
-                    # if error, we hide anyway
-                    isHidden = True
-                hidewhens_results.append(["%s/%s" % (formid, hwid), isHidden, hidewhen.isResetOnHide])
-        else:
-            hidewhens = self.getHidewhenFormulas()
-            if doc is None:
-                doc = getTemporaryDocument(db, self, REQUEST, validation_mode=True)
-            for hidewhen in hidewhens:
-                try:
-                    isHidden = self.runFormulaScript(
-                        SCRIPT_ID_DELIMITER.join(
-                            ['hidewhen', self.id, hidewhen.id, 'formula']
-                        ),
-                        doc,
-                        hidewhen.formula
-                    )
-                except PlominoScriptException, e:
-                    e.reportError(
-                        '%s hide-when formula failed' % hidewhen.id)
-                    # if error, we hide anyway
-                    isHidden = True
-                hidewhens_results.append([hidewhen.id, isHidden, hidewhen.isResetOnHide])
+        hidewhens = self.getHidewhenFormulas()
+        #if doc is None:
+        #    doc = getTemporaryDocument(db, self, REQUEST, validation_mode=True)
+        assert doc is not None
+        for hidewhen in hidewhens:
+            hidewhens_results.append(self._get_hidewhen(doc, hidewhen))
 
         # Set hidewhen values for multipage based on current page
-        if self.getIsMulti() and include_multipage:
+        if self.getIsMulti():
             num_pages = self._get_num_pages()
             current_page = self._get_current_page()
             # 0-indexed pages are ugly
@@ -1680,13 +1696,33 @@ class PlominoForm(Container):
         db.setRequestCache(cache_key, hidewhens_results)
         return hidewhens_results
 
+
+    security.declarePrivate('_get_hidewhen')
+    def _get_hidewhen(self, doc, hidewhen):
+
+        try:
+            isHidden = self.runFormulaScript(
+                SCRIPT_ID_DELIMITER.join(
+                    ['hidewhen', self.id, hidewhen.id, 'formula']
+                ),
+                doc,
+                hidewhen.formula
+            )
+        except PlominoScriptException, e:
+            e.reportError(
+                '%s hide-when formula failed' % hidewhen.id)
+            # if error, we hide anyway
+            isHidden = True
+        return [hidewhen.id, isHidden, hidewhen.isResetOnHide]
+
+
     security.declarePrivate('_get_hidden_subforms')
 
     def _get_hidden_subforms(self, REQUEST, doc, already_hidden=False, validation_mode=False):
         # print "in _get_hidden_subforms for %s" % self.id
         db = self.getParentDatabase()
         hidden_forms = []
-        hidewhens = self._get_hidewhens(REQUEST, doc, include_multipage=True)
+        hidewhens = self._get_hidewhens(doc)
         html_content = self._get_html_content()
         # If we're already hidden, any subforms are also hidden
         if already_hidden:
@@ -1724,7 +1760,7 @@ class PlominoForm(Container):
         db = self.getParentDatabase()
         hidden_fields = []
         reset_fields = []
-        hidewhens = self._get_hidewhens(REQUEST, doc, include_multipage=True)
+        hidewhens = self._get_hidewhens(doc)
         html_content = self._get_html_content()
         for hidewhenName, doit, isResetOnHide in hidewhens:
             if not doit:  # Only consider True hidewhens
@@ -1933,8 +1969,8 @@ class PlominoForm(Container):
         Check if any of those types are present.
         """
         db = self.getParentDatabase()
-        cache_key = 'hasFieldTypes_%d_%d_%d' % (
-            hash(self),
+        cache_key = 'hasFieldTypes_%s_%d_%d' % (
+            repr(self),
             hash(types),
             hash(applyhidewhen)
         )
@@ -2213,9 +2249,10 @@ class PlominoForm(Container):
             validation_mode=True,
             request=REQUEST)
 
-        hidden_fields, _ = self._get_hidden_fields(REQUEST, doc)
+        #getFormFields above doesn't remove hidden fields
+        hidden_fields, _ = self._get_hidden_fields(REQUEST, tmp)
 
-        hidden_forms = self._get_hidden_subforms(REQUEST, doc)
+        hidden_forms = self._get_hidden_subforms(REQUEST, tmp)
         for form_id in hidden_forms:
             form = db.getForm(form_id)
             if form:
@@ -2229,6 +2266,7 @@ class PlominoForm(Container):
 
         fields = [field for field in fields
                   if field.getId() not in hidden_fields]
+
         errors = []
         for f in fields:
             field_errors = []