+++ CHANGES.rst
@@ -91,6 +91,9 @@ Changelog
 - Call to unindexing document after being indeleted regardless it is deleted with Plone API or not [quang]
 - Fix regression with form onDisplay to accept different type of returned value [quang]
 
+- cache and optmise calls to validation and hidewhen formulas and temporary document creation during form load, and next page. Further optimisaiton on save required. [djay] 
+
+
 2.0b4 (2016-05-27)
 ------------------